Utilizing JPA's EntityManager ... I have to comment @PersistenceContext(unitName = "AffableBeanPU")     otherwise I get a deployment error     "Could not resolve a persistence unit corresponding to the persistence-context-ref-name" ... reconciled minor difference between MV and BV https://netbeans.org/kb/docs/javaee/ecommerce/transaction.html#entityManager

+++ sandbox/affableBean/src/java/controller/ControllerServlet.java
@@ -171,7 +171,7 @@ public class ControllerServlet extends HttpServlet {
         } else if (userPath.equals("/updateCart")) {
             
             int productId = Integer.parseInt(request.getParameter("productId"));
-            int quantity = Integer.parseInt(request.getParameter("quantity"));
+            short quantity = Short.parseShort(request.getParameter("quantity"));
             cart.update(productId, quantity);
             
             userPath = "/cart";
